requirements.txt
anthropic
python-dotenv
OpenAI